Create processNetworkFiles.py
It processes pcap files (generated by tcpdump) to compute the number of packages and bytes transmitted over the network. It generates a CSV file containing for all pcap files in a folder the number of packages and bytes transmitted. Note that the Python script networkMemoryAndCPU obtains pcap files associated to Android apps while playing a scenario.